Resources and resource groups either named or tagged will be culled
Previously there was an issue where if a resource group was tagged 'delete:1' it would be ignored

This was because only resources were checked if they were tagged. The same was also true for if a resource was named 'delete-me'.

It is now even more important to make sure that you haven't got a resource called do-not-delete :)